backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
psycopg2-binary==2.9.6
boto3==1.26.137
python-multipart==0.0.6
pydantic==1.10.7
requests==2.30.0